Added support for help functionality.
Help is now available with the optional secondary argument. All listed commands are summarised when help is called with no secondary argument. Added checks for number of secondary arguments that each completed commands have and sends the appropriate error message to the user when appropriate. Fixed naming convention from msg to message.