Fixed caps lock issue in kick command
Command handling now can receive arguments without changing caps. The parts that are not case sensitive are properly handled accordingly (especially the firstArgument). Error handling for missing send messages permissions is now properly handled. Permission requirements when the bot is missing some is now sent to the channel to remind the user to give the appropriate permission(s)/alternative approach. Help.js is now sent as a text message to the text/dm channels instead of embedded message.