Better error handling in kick.js and added statuslog.js functionality
Error handler in kick.js now can account for being unable to view the server as well as inability to send messages to the logging channel. Made logs for kicking user now prettier. statuslog.js can now print the appropriate channel redirect given it exists. May need to add a check for channel existence in the cache for statuslog and reset if necessary.